==> 03-cleaning.Rmd <==
# Data transformation

```{r, message=FALSE, warning=FALSE, echo=FALSE}
library(googlesheets4)
library(readr)
#library(blsAPI)
library(rjson) 
library(tidyverse)
library(zoo)
# library(lubridate)
```

## CPI data
The cpi data is accessed through cdc website api via ```googlesheets4::read_sheet```. We firstly add the item names to our main data of interest by matching corresponding seriesID. Since all attributes are in the string format, it's helpful to convert the item names into the factor type since we are interested in comparing time series for different items and cpi values into numeric values. We also create a new date attribute by concetenating year and month to get the full range of time series, which is further converted into the yearmon type. Finally, we create another data ```cpi_monthly.csv``` by ```pivot_wider``` to obtain monthly cpi values for all item names so that the number of observations are equivalent to the number of distinct year-months. This data will be used to merge with covid data below.


```{r, message=FALSE, warning=FALSE, echo=FALSE}
gs4_deauth()
cpi = googlesheets4::read_sheet('1fpEQ4wuuDlnfKokjKyRUrOcIwHUnIBmnBFubCUpnQ94',sheet='chained_data')
cpi1= googlesheets4::read_sheet('1fpEQ4wuuDlnfKokjKyRUrOcIwHUnIBmnBFubCUpnQ94',sheet='Sheet1')
chained = googlesheets4::read_sheet('1fpEQ4wuuDlnfKokjKyRUrOcIwHUnIBmnBFubCUpnQ94',sheet='chained')
```


```{r, message=FALSE, warning=FALSE, echo=FALSE}
cpi$item_name <- chained$item_name[match(cpi$seriesID, chained$seriesid)]
cpi$item_name = as.factor(cpi$item_name)
cpi$value = as.numeric(cpi$value)
cpi$date <- paste(cpi$year, "-", substr(cpi$period,2,3), sep="")
cpi$date <- as.yearmon(cpi$date, "%Y-%m")
#write.csv(cpi,"data/cpidata.csv", row.names = FALSE)
```

```{r, message=FALSE, warning=FALSE, echo=FALSE}
head(cpi)
```

### Travel
For the monthly transportation statistics, we only extract the ```Date``` and ```U.S. Airline Traffic - Total - Seasonally Adjusted``` for our primary interest in transportation, where we convert the dates into the ```yearmon``` type and remove all ```NA```s so that we are left with air passengers from Jan 2017 to Jul 2021.

```{r}
Monthly_Transportation_Statistics <- read_csv("data/Monthly_Transportation_Statistics.csv")
head(Monthly_Transportation_Statistics)
```


```{r}
Monthly_Transportation_Statistics <- read_csv("data/Monthly_Transportation_Statistics.csv")
Monthly_Transportation_Statistics = Monthly_Transportation_Statistics %>% select(travel = "U.S. Airline Traffic - Total - Seasonally Adjusted", date = "Date")

Monthly_Transportation_Statistics = Monthly_Transportation_Statistics %>% drop_na()
Monthly_Transportation_Statistics$date = as.Date(substr(Monthly_Transportation_Statistics$date, 1, 10), format = "%m/%d/%Y")

write.csv(Monthly_Transportation_Statistics,"data/travel.csv", row.names = F)
```



### Oil 

For oil data, we simply remove the observations prior to Jan 2017 to match the transportation data above and convert the dates into the ```yearmon``` type as discussed in section 2.


```{r}
oil <- read_csv("data/U.S._Field_Production_of_Crude_Oil.csv")
oil$Month = as.Date(oil$Month)
oil = oil %>% filter(Month >= "2017-01-01")
write.csv(oil,"data/oil.csv", row.names = F)
```

## Covid data

### United States COVID-19 Cases and Deaths by State overTime

Observe that covid data are daily based and cpi data are monthly based, we use ```dplyr``` to shrink covid data to monthly based by summing the number of cases and deaths for every state, where we denote this new dataset ```cdcmonth.csv``` after adjusting date into the ```yearmon``` format. We also noticed that for the attributes ```consent_cases``` and```consent_deaths``` some ```NA```s are denoted as the string ```'N\A'```, where it makes sense to replace them with actual ```NA``` since they are indeed missing values. In addition, we want to investigate cross state covid cases, where we want to focus on 50 main states. In particular, we have 60 states in total and the reported observations for New York City and Washington D.C. are separated from New York and Maryland respectively, where other states are disjoint from the 50. Hence, we adjust the observations for New York City and Washington D.C. and add them to New York and Maryland in ```cdcmonth.csv``` respectively. Furthermore, we replace the ```NA```s in probable cases and deaths with 0s and further replace confirmed cases and deaths with total cases - probable cases and total deaths - probable deaths since some of the confirmed cases and deaths are < 0 as mentioned in section 2 and probable cases are considered to be unreliable records. We have also created 3 new datasets ```cdc_all.csv```, ```cdc_monthly_all_state.csv``` and ```data_monthly.csv```, where ```cdc_all.csv``` is obtained from getting the daily reported cases and deaths for the entire US by summing over the cases and death across all the states, then we further shrink the data to the national monthly cases and deaths ```cdc_monthly_all_state.csv``` by grouping the data by each month. We eventually merge ```cdc_monthly_all_state.csv``` and ```cpi_monthly.csv``` obtained above to conduct analysis between covid and cpi attributes as well as performing visualizations such as plotting us maps.


```{r, message=FALSE, warning=FALSE, echo=FALSE}
cdc = read_csv("data/United_States_COVID-19_Cases_and_Deaths_by_State_over_Time.csv")
```


```{r, message=FALSE, warning=FALSE, echo=FALSE}
cdc[,c('consent_cases')][cdc[,c('consent_cases')] == 'N/A'] = NA
cdc[,c('consent_deaths')][cdc[,c('consent_deaths')] == 'N/A'] = NA
```


```{r, message=FALSE, warning=FALSE, echo=FALSE}
cdc$date = as.Date(cdc$submission_date, '%m/%d/%Y')
cdc$month = as.yearmon(format(as.Date(cdc$date), "%Y-%m"))
```


```{r, message=FALSE, warning=FALSE, echo=FALSE}
cdc.month = cdc %>% mutate(state.mon = paste(state, as.character(month), sep='') ) %>% group_by(state.mon) %>% summarize(new_case = sum(new_case, na.rm = TRUE), new_death = sum(new_death, na.rm = TRUE), month = min(month), state=min(state)) %>% rename(date = month) %>% select(-c(state.mon))
```

```{r, message=FALSE, warning=FALSE, echo=FALSE}
cdc.month$date = as.character(cdc.month$date)
cdc.month$date = paste(substr(cdc.month$date,5,8), match(substr(cdc.month$date,1,3), month.abb), sep="-")
```

```{r merge_city_to_state}
cdc.month$statefull = state.name[match(cdc.month$state,state.abb)]
cdc.month$state[cdc.month$state=="DC"] = "MD"
cdc.month$state[cdc.month$state=="NYC"] = "NY"
cdc.month$statefull[cdc.month$state=="DC"] = "Maryland"
cdc.month$statefull[cdc.month$state=="NYC"] = "New York"
```


```{r}
head(cdc.month)
```


```{r}
write.csv(cdc.month,"data/cdcmonth.csv", row.names = FALSE)

```


```{r}
cdc = cdc  %>% mutate(prob_cases = ifelse(is.na(prob_cases), 0, prob_cases)) %>% mutate(conf_cases = tot_cases - prob_cases)
cdc = cdc  %>% mutate(prob_death = ifelse(is.na(prob_death), 0, prob_death)) %>% mutate(conf_death = tot_death - prob_death)
cdc = cdc %>% mutate(date=as.Date(date))
write.csv(cdc,"data/cdc.csv", row.names = FALSE)
```


```{r}
cdc = read.csv("data/cdc.csv")
cdc = cdc %>% mutate(date=as.Date(date), month=as.yearmon(format(as.Date(date), "%Y-%m")))

### total death working
cdc_all = cdc %>% select(date, month, new_death, new_case, conf_cases, conf_death) %>% 
    group_by(date) %>% 
    summarise(month=max(month), new_death=sum(new_death),new_case=sum(new_case), conf_cases=sum(conf_cases), tot_death=sum(conf_death))
# calculate monthly data
cdc_monthly_all_state = cdc_all %>% 
  select(month, new_death, new_case, conf_cases, tot_death) %>% 
  group_by(month) %>% 
  summarise(new_death=sum(new_death),new_case=sum(new_case), tot_cases=max(conf_cases), tot_death=max(tot_death))

# combine cpi and cdc
selection = c("All items", "Energy","Apparel", "Commodities", "Durables", "Education", "Communication", "Food", "Housing", "Fuels and utilities", "Medical care commodities", "Medical care services", "Recreation", "Transportation", "Private transportation", "Public transportation" )
cpi = cpi %>% filter(item_name %in% selection)
cpi_monthly = cpi %>% select(date, value, item_name) %>% pivot_wider(id_cols = , names_from = item_name, values_from = value)
data_monthly = left_join(cpi_monthly, cdc_monthly_all_state, by=c("date"="month")) %>% rename(month = date)

#
write.csv(cdc_all,"data/cdc_all.csv", row.names = F, fileEncoding="UTF-8")
write.csv(cdc_monthly_all_state,"data/cdc_monthly_all_state.csv", row.names = F, fileEncoding="UTF-8")
write.csv(cpi_monthly,"data/cpi_monthly.csv", row.names = F, fileEncoding="UTF-8")
write.csv(data_monthly,"data/data_monthly.csv", row.names = F, fileEncoding="UTF-8")
```

```{r, message=FALSE, warning=FALSE, echo=FALSE}
cdc.month = read.csv("data/cdcmonth.csv")
cdc.month$date <- as.yearmon(cdc.month$date, "%Y-%m")
```


```{r}
head(cdc.month)
```
