==> 05-results.Rmd <==
# Results

```{r packages, message=FALSE, warning=FALSE, echo=FALSE}
library(googlesheets4)
library(rjson) 
library(tidyverse)
library(graphics)
library("readxl")
library("lhs")
library(splines)
library(maps)
library(mapdata)
library("viridis")
library(RColorBrewer)
library(zoo)
# library(lubridate)
library(highcharter)
library(patchwork)
```

```{r, message=FALSE, warning=FALSE, echo=FALSE}
cdc = read.csv("data/cdc.csv")
cdc_all = read.csv("data/cdc_all.csv", fileEncoding="UTF-8")
cdc_monthly_all_state = read.csv("data/cdc_monthly_all_state.csv", fileEncoding="UTF-8")
cpi_monthly = read.csv("data/cpi_monthly.csv", fileEncoding="UTF-8")
data_monthly = read.csv("data/data_monthly.csv", fileEncoding="UTF-8")
cdc.month = read.csv("data/cdcmonth.csv")
cpi = read.csv("data/cpidata.csv")
oil = read_csv("data/oil.csv")
travel = read_csv("data/travel.csv")
```

```{r, message=FALSE, warning=FALSE, echo=FALSE}
cdc = cdc %>% mutate(date=as.Date(date), month=as.yearmon(format(as.Date(date), "%Y-%m")))

cdc_all = cdc_all %>% mutate(date=as.Date(date), month=as.yearmon(format(as.Date(date), "%Y-%m")))

cdc_monthly_all_state = cdc_monthly_all_state %>% mutate(month=as.yearmon(month))

cpi_monthly = cpi_monthly %>% mutate(date=as.yearmon(date))

data_monthly = data_monthly %>% mutate(month=as.yearmon(month))

cdc.month = cdc.month %>% 
  mutate(date=as.yearmon(date, "%Y-%m")) %>% 
  group_by(statefull,date) %>% 
  dplyr::summarise(new_case=sum(new_case),new_death=sum(new_death))

cdc_monthly_by_state = cdc %>% 
  select(month, state, new_death, new_case, conf_cases, tot_death) %>% 
  group_by(month, state) %>% 
  summarise(new_death=sum(new_death),new_case=sum(new_case), tot_cases=max(conf_cases), tot_death=max(tot_death)) %>% 
  mutate(statefull=state.name[match(state,state.abb)]) %>% 
  drop_na()

selection = c("All items", "Energy","Apparel", "Commodities", "Durables", "Education", "Communication", "Food", "Housing", "Fuels and utilities", "Medical care commodities", "Medical care services", "Recreation", "Transportation", "Private transportation", "Public transportation" )
cpi = cpi %>% mutate(date = as.yearmon(date)) %>% filter(item_name %in% selection)

colnames(oil)  = c('Month','oil')
```

## Covid Development in U.S.

### Cases Trend Line

The spread of covid was extremely fast such that people's lives were significantly impacted all over the nation. The cumulative Covid cases in U.S. have currently surpassed 40 million. The first two plots will demonstrate how Covid has been developing since the time it started.

#### Cumulative Total Cases
```{r covid develop trend}
ggplot(cdc_all,
       aes(x=date, y=conf_cases/1000)) +
  geom_area(alpha=0.8, fill="#0d2b53") + 
   scale_x_date(date_labels = "%m-%d",
               date_breaks= "4 week" 
               ) + 
  labs(x = "Date", y = "Cases Number (in thousands)", caption = "Source: CDC", 
       title = "Covid Development Trend in U.S." ,subtitle = "Cumulative Total Confirmed Cases"
       ) +
  theme_bw() +
  theme(plot.title = element_text(face = "bold"),
        plot.subtitle = element_text(face = "bold", color = "grey35"),
        plot.caption = element_text(color = "grey68"),
        axis.text.x = element_text(angle = 45, hjust = 1),
        legend.position="bottom")
```

Covid started in March 2020 and reaches highest increase rate during winter 2020. In the first half of the year, covid cases number had surged to 5 million from 0, and it further increased to 25 million in the next half year. The increase of Covid cases was in an exponential rate as people went outdoors more frequently to socialize since the spread of the coronavirus also posseseses 'Network Effect'. 

#### Daily New Cases
```{r New Cases Trend Line, message=FALSE}
ggplot(cdc_all,
       aes(x=date, y=new_case/1000)) +
  #geom_point() + 
  geom_area(alpha=0.5, fill="#2a6998") + 
  #geom_line(stat = "smooth", method = "loess", span=0.3, size = 1.5, color="firebrick2") + 
  geom_smooth(method = "loess", span=0.3, size = 1.5, color='#95b5f0', se=F) +
  scale_x_date(date_labels = "%m-%d",
               date_breaks= "4 week" #,limits=c(as.Date('2020-03-01'), max(cdc_all$date))
               ) + 
  #scale_y_continuous(limits=c(0,15)) +
  scale_color_manual(values=c("dodgerblue3", "firebrick2")) + 
  labs(x = "Date", y = "Cases Number (in thousands)", caption = "Source: CDC", 
       title = "Covid Development Trend in U.S." ,subtitle = "Daily New Cases"
       ) +
  theme_bw() +
  theme(plot.title = element_text(face = "bold"),
        plot.subtitle = element_text(face = "bold", color = "grey35"),
        plot.caption = element_text(color = "grey68"),
        axis.text.x = element_text(angle = 45, hjust = 1),
        legend.position="bottom")

```

The daily new cases plot further justified what we mentioned above. The winter holiday in 2020 created an opening for facilitation of the spread of the virus in the United States, where the cold weather also enabled the Virus to live longer in the environment. In contrast, with the aid of a warmer weather, promoting and limiting social activities afterwards played a vital role in the obstruction of the spread during summer 2021, where we can observe that numbers of new cases are a lot less. 

## Covid Development in Different States

Here we compared how new cases and total cases change from 1976 to 2020 by state so as to see how Covid development differs in different location. 

Instructions to use the graph:

* Play/pause clicking the play/pause button

* Navigate the motion slider by dragging the slider thumb

* Navigate the motion slider by hitting the left and right arrow keys.

```{r new cases}
cdcmonthcase = cdc.month %>% group_by(statefull)%>%arrange(date)%>% 
    do(item = list(
        state = first(.$statefull),
        sequence = .$new_case,
        value = first(.$new_case)
    ))%>%
    .$item

highchart(type = "map") %>% 
  hc_add_series(data = cdcmonthcase,
                name = "Dem. Margin",
                mapData = usgeojson,
                joinBy = c("name","state"),
                borderWidth = 0.05,
                dataLabels = list(enabled = TRUE, format = "{point.properties.postalcode}")
                ) %>% 
  hc_colorAxis(minColor = "white", maxColor = "orange") %>%
  hc_title(text = "new cases from 2020 Apr to 2021 Nov by state") %>%
  hc_add_theme(hc_theme_smpl()) %>%
  hc_motion(
    enabled = TRUE,
    axisLabel = "yearmonth",
    labels = as.character(sort(unique(cdc.month$date),decreasing = FALSE))
  )%>% 
  hc_legend(title = list(text = "new cases"), valueDecimals = 0, valueSuffix = "%") %>%
   hc_mapNavigation(enabled = TRUE) %>% hc_exporting(enabled = TRUE)
```

From the motion map, we can see that the first five cases in the United States come from Illinois, Washington and Arizona. At the beginning, California and New York have the highest number of new cases, which are two of the largest states that have the most population. From Jun 2020 to Sep 2020, it turned out that California, Texas and Florida continuously had the highest number new cases, followed by New York in a later period. A reasonable explanation is that population for these states are sufficiently large to enable the efficient spread of Coronavirus virus, especially if the restrictions are not rigorous enough.

```{r tot cases}
cdcmonthtotal = cdc_monthly_by_state %>% 
  group_by(statefull)%>%
  arrange(month)%>% 
    do(item = list(
        state = first(.$statefull),
        sequence = .$tot_cases,
        value = first(.$tot_cases)
    ))%>%
    .$item

highchart(type = "map") %>% 
  hc_add_series(data = cdcmonthtotal,
                name = "Dem. Margin",
                mapData = usgeojson,
                joinBy = c("name","state"),
                borderWidth = 0.05,
                dataLabels = list(enabled = TRUE, format = "{point.properties.postalcode}")
                ) %>% 
  hc_colorAxis(minColor = "white", maxColor = "orange") %>%
  hc_title(text = "total cases from 2020 Jan to 2021 Nov by state") %>%
  hc_add_theme(hc_theme_smpl()) %>%
  hc_motion(
    enabled = TRUE,
    axisLabel = "yearmonth",
    labels = as.character(sort(unique(cdc.month$date),decreasing = FALSE))
  )%>% 
  hc_legend(title = list(text = "total cases"), valueDecimals = 0, valueSuffix = "%") %>%
   hc_mapNavigation(enabled = TRUE) %>% hc_exporting(enabled = TRUE)
```

From the motion map, we can observe that New York is the state with the largest total cases in the beginning and then from Jul 2020 until Nov 2021, California, Texas and Floria are the three States that have the largest total cases which is consistent with the trend of new cases from the previous motion map. Again, these are states with large populations and decent weather acts like a catalyst for the virus to live and spread even more rapidly.

## Clustering States by Covid Trend


```{r spline map, message=FALSE, warning=FALSE, echo=FALSE}
source('alvin-clustering.r')
for (i in 9) {
# Run K-means to cluster the knots of the states
  # Bcoef.mat = plot.bspline(covid.com, all.states, 'NEWCD_NORM_500', n.knots=15, degree=3)
  
  # choose the number of clusters we want and assign each state its closest cluster
  set.seed(i)

  for (k in 6) {
    cluster.num = k
    clusters = kmeans(as.data.frame(Bcoef.mat), cluster.num, nstart = 10)$cluster
    sort.cluster.list = list()
    for (j in 1:k) {
      sort.cluster.list[[j]] = -1
    }
    
    sort.clusters = sort.state.clusters(clusters, sort.cluster.list)

    states.df = data.frame(state = all.states, CLUSTER = sort.clusters)
    # write.csv(states.df, paste("./clusterings/cluster", i, ".csv", sep=""))
    
    state.cluster = list()
    for (i in 1:cluster.num) {
      state.cluster[[i]] = subset(states.df, CLUSTER==i)$state
    }
    cs = ifelse (map.df$state %in% state.cluster[[cluster.num]], cluster.num, cluster.num+1)
    for (i in (cluster.num-1):1) {
      cs = ifelse (map.df$state %in% state.cluster[[i]], i, cs)
    }
    
    map.df$cluster = cs
    #map.df$cluster = ifelse (map.df$STATE %in% state.cluster[[1]], 1,
    #                 ifelse (map.df$STATE %in% state.cluster[[2]], 2,
    #                 ifelse (map.df$STATE %in% state.cluster[[3]], 3,
    #                 ifelse (map.df$STATE %in% state.cluster[[4]], 4,
    #                 ifelse (map.df$STATE %in% state.cluster[[5]], 5,
    #                 ifelse (map.df$STATE %in% state.cluster[[6]], 6, 7
    #                        ))))))
    
    # us map on bspline knots clusters
    colors = c("#BBE7FE", "#D3B5E5","#FFED8A", "#FFD4DB", "#EFF1DB", "darkseagreen2")
    us.bspline.map = ggplot(map.df, aes(x=long, y=lat, group = group)) + geom_polygon(colour="black", size=0.5)
    for (i in 1:cluster.num) {
      us.bspline.map = us.bspline.map + geom_polygon(fill=colors[i], data = filter(map.df, cluster == i))
    }
    print(us.bspline.map + map.state.names + ggtitle(paste('U.S. Map spline knots kmeans covid cases k=', k, sep='')))
  }
}
```

To identify difference and similarities between Covid case curves, we estimated a basis spline (B-spline) model for every state. Each estimated B-spline is a weighted piecewise combination of 10 polynomials, connected at the “knots”. The estimated splines in terms of their weight coefficients are closely related in states with similar underlying case curves. We compared the estimated state splines by using the K-means algorithm(We set k = 6) to cluster similar sets of weight coefficients, identifying groups of states with similar Covid case trajectories. In general, we can observe that the coefficients for two states in the same cluster are more similar than those in a different cluster(e.g. Florida and Texas). (Relevant link: https://towardsdatascience.com/using-b-splines-and-k-means-to-cluster-time-series-16468f588ea6)

```{r covid spline cases, message=FALSE, warning=FALSE, echo=FALSE}
for (i in c(2,4,6)) {
  print(plot_states(cdc.month, state.cluster[[i]], "new_case", "Covid Monthly New Cases Time Series"))
}
```

The plotted trend line further accentuated that the clustering model is a reasonable method to group the cases curves across all states. For instance, `California`, `Florida` and `Texas` are grouped, which is consistent with the observation from interactive US map above. Those states are in the southern area of America, implying that a warmer weather that provides a fit environment for the virus to boost its longevity and activity. 


## Consumer Price Index Development

We utilized chained consumer price index to reflect the change in our living expenses. We aim to find the item that had highest increase in price after the covid breakout. We will also analyse their similarities with covid case patterns.

### Chang of Living Expenses by Items

```{r}
itemname = levels(as.factor(cpi$item_name))
DF=data.frame()
for (i in itemname){
  v = cpi %>% select(item_name,date,value) %>% filter(item_name == i,date == as.yearmon("2020-01"))
  v = v$value
  b = cpi %>% select(item_name,date,value) %>% filter(item_name == i,date == as.yearmon("2019-01"))
  b = b$value
  temp = cpi %>% select(item_name,date,value) %>% filter(item_name == i,date == as.yearmon("2021-10"))
  t =  cpi %>% select(item_name,date,value) %>% filter(item_name == i,date == as.yearmon("2019-12"))
  t = (t$value - b)/b
  temp$rateafter = (temp$value - v)/v
  temp$ratebefore = t
  DF = rbind(DF,temp)
}
c = select(DF,c("item_name","ratebefore","rateafter"))
c =c %>% gather("type", "value", -item_name)
```

```{r index percent change}
ggplot(c, aes(y=reorder(item_name,value),x = value * 100,fill = type))+
  geom_bar(stat="identity", position = position_dodge())+
  theme(axis.text.x = element_text(angle = 90))+
  labs(x = "Percentage Change", y = "CPI Items", 
       title = "Price Index Percentage Changes" ,
       subtitle = "Compare Price Index Change Rate in the Year Before and After Covid",
       fill = "Change Rate Period"
  ) +
  #scale_fill_brewer(palette = "BuPu")+
  scale_fill_hue(labels = c("After Covid", "Before Covid"))+
  theme_bw()
```

From the bar plot, we can see that public transportation is the only one that decreases the CPI after the Covid-19. Because of the Covid-19, many people avoided using `pubilic transportation` which is crowded. The decrease of demands decreased the CPI. In the opposite, `private transportation`, `Energy` and `transportation` CPI increases a lot after the Covid-19 comparing with the change rates before the Covid-19. More people choose to use `private transportation` to reduce the chance of exposure to the crowds. We already know that energy is a necessity for `private transportation`. The increase of demands on these three products increases the CPI at the same time. Durable goods is the one that has negative change in CPI before the Covid-19 and then increases sharply after the Covid-19. We know that during the Covid-19, people are forced to stay at home and this may be the reason that people bought lots of durable goods as they spent less outdoor. Moreover, the increase in `Energy` led to the manufacturing and transportation cost to increase which leads to a higher price of goods in the store. 

### All Items' Price Trend

Here we put all CPI items together for display so that we can know how the trend look like. The graph visualization is not idea and may not be informative. We will solve the problem and apply scaling, grouping, and correlation analysis on those data so as to draw useful insights and clear graphs. 

```{r cpi trends, fig.height = 6, message=FALSE, warning=FALSE, echo=FALSE}
cpi %>%
  ggplot( aes(x=date, y=value, group=item_name, color=item_name)) +
  geom_line() + ggtitle("CPI Trend Lines for All Items") + 
  theme_classic()+
  labs(colour="CPI Items", y="CPI", x='Month')+
  theme(axis.text.x = element_text(angle = 0),legend.text = element_text(size=5),legend.position = 'right')
```

### Correlation Between Consumer Items
From the first graph with all CPI items, it is hard for us to tell clearly how each one moves. And by their names and our common sense, it is very likely that some of them share similar trend. Therefore, we decided to use a heat map to represent their correlations so that we can evaluate their similarities quantitatively. Our purpose here is to find similarities and group similar ones so that we can find the main cause of higher prices easier. 

```{r cor mat}
cor_matrix = cpi_monthly %>%  select(-date) %>% cor() %>% as.data.frame()
source('hchart _corr.R')
hchart.cor(cor_matrix)
```

According to the heat map, we can easily tell that many of the CPI items are highly and positively correlated with each other. However, `Public Transportation`, `Medical Care Commodities`, `Communication`, and `Apparel` showed white and red color in the heat map. This indicates that they are not or negatively correlated with other items. Combining with the trend graph, I know that their prices did not increase much. It is also reasonable because some of them have pricing mechanism that is not determined by the market. For instance, `Public Transportation`, `Medical Care Commodities` and `Communication` have government set prices and fixed prices.

`Energy` are highly correlated with some other industrial and transportation related items. According to the heatmap, I would group the following items together: `Energy`,  `Commodities`, `Transportation`, `Private Transportation`, `Durables` and `Fuel and Utilities`. 

The rest of items are `Food`, `Education`, `Housing`, `Medical Care Services` and `Recreation`. Their correlation are higher than 0.8 indicating a high correlation between them. Their prices are highly affected by labor cost in services area. Since our living expenses increased, people working at services industry would need higher salaries and results in higher price for services. Moreover, their correlation with `Energy` is relatively low. Just like I explained before, it would take time to have effect on those prices. People would first feel the living expense to increase before they ask for higher salary. The delaying effect caused the correlation between them and `Energy` to be low. 

Therefore, we grouped similar CPI items together and will show their similarities in the next part. 

### Group CPI Items

The groups we created for CPI items are as follows:
1. **Price Stable Items**: `Public Transportation`, `Medical Care Commodities`, `Communication`, and `Apparel`
2. **Energy Related Items**: `Energy`,  `Commodities`, `Transportation`, `Private Transportation`, `Durables` and `Fuel and Utilities`
3. **Labor Related Goods**: `Food`, `Education`, `Housing`, `Medical Care Services` and `Recreation`

In order to confirm their similarities, their trend lines are plotted together below. Since their base values are different, the difference in scale would cause some trends not as obvious as others with larger scale. The standardized method is used here to showcase the similarities better. 

```{r cpi group similarities,message=FALSE, warning=FALSE, echo=FALSE}
data_monthly_sclae_plot = data_monthly %>% 
  mutate(across(All.items:tot_death,scale)) %>% 
  pivot_longer(cols = All.items:tot_death, names_to='var',values_to='value') 

selection = c('Public.transportation','Medical.care.commodities', 'Communication', 'Apparel')
p1 = data_monthly_sclae_plot %>% filter(var %in% selection) %>% 
  mutate(value=scale(value)) %>% 
  ggplot(aes(x=month, y=value, color=var)) + 
  geom_line(alpha=0.4)+
  geom_smooth(method = 'loess', span=0.5,se=F)+
  geom_point(alpha=0.4)+
  ggtitle("Scaled CPI Trend")+
  labs(colour="CPI Items", colour="CPI Items", y="Scaled CPI", x='Month',subtitle = "Stable Price Group")

selection = c('Energy','Commodities', 'Private.transportation', 'Durables')
p2 = data_monthly_sclae_plot %>% filter(var %in% selection) %>% 
  mutate(value=scale(value)) %>% 
  ggplot(aes(x=month, y=value, color=var)) + 
  geom_line(alpha=0.4)+
  geom_smooth(method = 'loess', span=0.5,se=F)+
  geom_point(alpha=0.4)+
  ggtitle("Scaled CPI Trend")+
  labs(colour="CPI Items", y="Scaled CPI", x='Month',subtitle = "Energy Related Group") #colour="CPI Items",

selection = c('Food','Education', 'Housing', '`Medical.care.services', 'Recreation')
p3 = data_monthly_sclae_plot %>% filter(var %in% selection) %>% 
  mutate(value=scale(value)) %>% 
  ggplot(aes(x=month, y=value, color=var)) + 
  geom_line(alpha=0.4)+
  geom_smooth(method = 'loess', span=0.5,se=F)+
  geom_point(alpha=0.4)+
  ggtitle("Scaled CPI Trend")+
  labs( y="Scaled CPI", x='Month',subtitle = "Labor Related Group")
selection = c('Education','Energy', 'Communication')
p4 = data_monthly_sclae_plot %>% filter(var %in% selection) %>% 
  mutate(value=scale(value)) %>% 
  ggplot(aes(x=month, y=value, color=var)) + 
  geom_line(alpha=0.4)+
  geom_smooth(method = 'loess', span=0.5,se=F)+
  geom_point(alpha=0.4)+
  ggtitle("Scaled CPI Trend")+
  labs(colour="CPI Items", y="Scaled CPI", x='Month',subtitle = "Comparison") 
p1 
p2 
p3
p4
```

From the scaled trend lines, it proves that our grouping did a pretty good job! **Price Stable Items** had price drop since early 2019 and started to rise in price after mid 2020. **Energy Related Items** had a increase trend in early 2019 and had a huge drop after Covid broke out and increase in price afterwards dramatically. **Labor Related Items** had a general increasing trend since 2019, and had a higher slop after Covid broke out. The last graph showed one item from each of three groups to showcase their difference in trend as response to the Covid development.

According to the trends of different groups, **Energy Related Items** are more interesting to us. Their price dropped a lot after Covid just broke out. And then, the price increased dramatically after the first several months. Other goods, doesn't seem to have such big impact from Covid. Therefore, we think **Energy Related Items** will be the target we will further investigate so as to discover how Covid caused our living expense to increase. 

## Covid's Impact on Energy Price Index

### Covid & Energy Trend Comparison

```{r Covid & Energy Trend Comparison, message=FALSE, warning=FALSE, echo=FALSE}
selection = c('tot_cases','Energy')
data_monthly_sclae_plot %>% filter(var %in% selection) %>% 
  mutate(value=scale(value)) %>% 
  ggplot(aes(x=month, y=value, color=var)) + 
  geom_line(alpha=0.4)+
  #geom_line(stat='smooth', method = 'loess', span=0.3,se=F, alpha=0.5, size=1)+
  geom_smooth(method = 'loess', span=0.3,se=F)+
  geom_point(alpha=1)+
  ggtitle("Covid & Energy Trend Lines")+
  labs(colour="Variable Name", y="Scaled Changed", x='Month',subtitle = "Showing How Energy Price Index Varies as Covid Develops")
```

We have scaled the data so as to make their changes obvious and comparable. At beginning of 2020, `Energy` started to drop from 0 which is its mean to over -1.5 times its standard deviation. At the end of 2020, though `Total Cases` was still growing, `Energy` price started to recover and increase to higher level than before Covid broke out. 


### Change of Energy Price Index

Next, we are going to investigate the probable reason that Covid could cause Energy price to change like what we discovered. We would discuss the cause from **supply** perspective and **demand** perspective. 

#### Demand Side Analysis
```{r travel change}
travel %>%
  hchart('line', hcaes(x = 'date', y = 'travel')) %>%
  hc_colors("grey")%>%
  hc_chart(zoomType = "x") %>%
  hc_legend(align = "right", verticalAlign = "top",layout = "horizontal",floating=TRUE,y=80,x=-100) %>%
  hc_xAxis(title = list(text = "date"),
           labels= list(style=list(fontSize='10px'),
                        rotation = -45, step = 1)) %>%
  hc_yAxis(title = list(text = "travel number")) %>%
  hc_title(text = "<b>change of travel from Jan 2017 to  Jul 2021 </b>") %>%
  hc_subtitle(text = "Click on the dot to show selected month; click on a dot and drag across to zoom in") %>%
  hc_exporting(enabled = TRUE)
```

Before the Covid-19, the travel number increase steadily until February 2020. At the begining of the Covid-19 spread out, from February 2020 to April 2020, the travel decreases sharply which follows the tread that CPI for energy decreases sharply from Jan 2020 to March 2020, and then CPI for energy rises a little in July 2020, stay until Dec 2020, and a slight decrease in Dec 2020. From the travel trend plot, we can also see that the travel number has a slow increase from April 2020 to February 2021. The CPI for energy and air travel number both increase from Jan 2021 until now.

Travel takes large porpotion of energy consumption. As the travel needs decrease, the demand to the energy decreased at the early 2020. The increase in travel in late 2020 and 2021 lead to the increase in energy demand. Thus, stimulates the Energy price to increase. 

It is clear that people stop traveling and working from home at the early stage of Covid broke out. This lead to a huge decrease in number of travel passengers number. In later 2020, the number of travellers is 1/3 of the number in early 2020. People starts to go back to office and go on vacation in summer 2021 as the new cases number starts to drop and people can not stand the lives staying at home everyday any more.

#### Supply Side Analysis
```{r oil supply}
oil %>%
  hchart('line', hcaes(x = 'Month', y = 'oil')) %>%
  hc_colors("grey")%>%
  hc_chart(zoomType = "x") %>%
  hc_legend(align = "right", verticalAlign = "top",layout = "horizontal",floating=TRUE,y=80,x=-100) %>%
  hc_xAxis(title = list(text = "date"),
           labels= list(style=list(fontSize='10px'),
                        rotation = -45, step = 1)) %>%
  hc_yAxis(title = list(text = "oil supply")) %>%
  hc_title(text = "<b>change of oil supply from Jan 2017 to  Sep 2021 </b>") %>%
  hc_subtitle(text = "Click on the dot to show selected month; click on a dot and drag across to zoom in") %>%
  hc_exporting(enabled = TRUE)
```

From the above line chart, we can see that before the covid-19, the supply of oil increases steadily. At the beginning of Covid broke out, from Mar 2020 to May 2020, the oil production decreases sharply which shares the same trend with travel number. The decrease of needs for travel leads to the decrease of needs for energy, results in  the decrease of energy price and finally cause the decrease of supply. Then the production increases slowly to a relatively steady level. From the CPI of Energy plot and the travel number plot, we can see that both of them increases from Jan 2021 until now. 

The increase of needs for travel increases the demand for energy. The extra needs for energy and the shortage of energy production together make the energy price index surges.
