==> 07-conclusion.Rmd <==
# Conclusion

## Limitations and Future Directions

From this work, we have concluded how COVID-19 pandemic impact on price of energy. In the analysis, we conclude that other items like `Food`, `Housing` and `Services` have delayed affect by the price increase of `Energy`. We assume that people are experiencing higher living expense due to `Energy` related price increase, so they ask for higher salary which yields to higher price in `Services`. The limitation is that we did not further analyse other reasons caused the price to increase. For example, in order to overcome the pandemic, U.S. government utilized monetary easing policy which will naturally cause inflation. Though the government policy is also highly related with Covid pandemic, it is the domain beyond the scope of this project. Further insights could be induced from analysis following that direction. 

In addition, when talking about the supply of energy, we only considered crude oil productions in the United States. Though USA is one of the top crude oil producers in the world, there are still many other countries such as Saudi Arabia, Russia and Canada which produce large amounts of crude oil as well. Besides, we focused on crude oil while electricity, gas, coal also have contribution to the energy price index, collecting these items' data and analyze their supply and demand might provide us with other insights. Moreover, we discovered that even though travel recovered in late 2021, the production of crude oil demonstrates a trivial sign of recovery. This might be due to the argument that higher price of crude oil leads to higher profit, but could also provide another direction for future discussions. 

In short, due to the various constraints, we made several assumptions and included the most significant change in our analysis. There are still many potential aspects that could be discussed to explain why would COVID-19 pandemic caused our living expenses to significantly surge. Thus, we are looking forward to deeper and more comprehensive analysis on this topic!

## Lessons Learned

From our project, we found that energy is a representative item which has high correlations with other items and COVID data. When investigating CPI change for energy, we found that the energy, air travel(demands of energy) and oil production(supply of energy) first decreased sharply as a result of COVID-19 because of the travel ban, which is followed by dilatory increase. However, despite the recovery of the air travel later on, oil production still remained to stay low. Hence, the gaps between demands and supply lead to an increase of energy CPI. Two possible reasons of absence of increase in the oil production are the incrementation of total revenue gained by company due to higher prices and the lack of labor due to the pandemic. There are definitely plenty of rooms for further investigations.

Throughout the project, we have achieved a better understanding of data visualization and analysis. In particular, we learned to use `API` to obtain data from Internet which we have never touched on before, dealt with unknowns, used regression methods like basis spline to conduct clustering that helps with visualisation. Finding the best way to present data will make a huge difference to what people can grasp and perceive. It is not only about making more beautiful and elegant graphs and pictures, but also inducing visual representations that are simultaneously intuitive and informative. Moreover, tuning the graph with labels, annotations, and combination of colors played an important role to make the plots more interpretable. Our passion in visualizing and analyzing data has significantly grown as we learn and work ourselves through this project, where the ultimate goal is to design graphic representations to reduce communication complexity so that majority of the population can comprehend our demonstrations in seconds.
